+++ CHANGELOG.md
@@ -7,6 +7,8 @@ Maplibre welcomes participation and contributions from everyone.
 - Android manifest cleanup for `libnavigation-android`
   - Explicitly add `android.permission.ACCESS_NETWORK_STATE`, as this is needed for `com.mapbox.services.android.core.connectivity.ConnectivityReceiver`
   - Remove Mapbox telemetry provider references (the references to code have already been removed)
+- Additional build version checks `if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU)` for deprecated context register receiver calls.
+- Added directions builder `waypointIndices` to NavigationRoute builder.
 
 ### v3.0.0 - November 5, 2023